chore(api): rename entity files
