+++ src/shared/entities/Base.ts
