+++ src/shared/entities/Session.ts
@@ -1,4 +1,4 @@
-import { BaseEntity } from "./BaseEntity";
+import { BaseEntity } from "./Base";
 
 export interface SessionEntity extends BaseEntity<"Session"> {
   userAgent: string;
