+++ src/shared/entities/index.ts
@@ -1,5 +1,5 @@
 export type EntityType = "Account" | "Session" | "User";
 
-export * from "./AccountEntity";
-export * from "./SessionEntity";
-export * from "./UserEntity";
+export * from "./Account";
+export * from "./Session";
+export * from "./User";
